Update test empty_string_host_t() for OTP 26
OTP-18543 which is introduced in OTP-26 states:
Looking up, connecting to and sending to a host with an empty name
is now handled by trying to look up the address of the root domain,
which fails. Previously some of these operations caused an internal
exception, which contradicted type specifications.

For details, see:
https://github.com/erlang/otp/commit/31a3022