Fix node reactivation issue
When a cluster is starting each node might have a different view of the slotmap.
Ered can first receive a slotmap which contains a specific node, and when
querying another node that specific node might not yet be known.
When a previously known node is missing from a received slotmap the client for
that node is deactivated, but not closed until the close_wait timeout.
If the specific node resurface in the next slotmap update then the peer is
reactivated.

1. This fixes a problem in `handle_info({slot_info..)` that State.closing still
   contained reactivated peers, NewClosing did not contain the result of start_clients/2.

2. node_is_available/2 could include nodes that are deactivated/closing.

3. Handle {connection_down, node_deactivated}.
   When deactivating a peer ered_cluster receives {connection_down, node_deactivated}.
   When a connected client, i.e. in up, was deactivated we previously removed it from
   the set `up`. If a connected peer is reactivated we dont get a connection_up, i.e.
   we would never add it to `up` again, we have a inconsistent state.
   If the connected peer was a master we would never get a cluster_ok.

This fixes intermittent issues seen in t_empty_initial_slotmap