Add a release CI workflow with a version check
Add a CI workflow that is only run when a tag is pushed to
Github, and a check that the ered application vsn is matching
the pushed tag.

application:get_key(ered, vsn) is used to get the vsn version
and the shell will return an error code when the version and
the tag is not matching, which fails the CI job.