API change for standalone mode (#129)
The `ered` module is the API for cluster and standalone mode.

For connecting, two different functions are used: `connect_cluster/2`
for cluster mode and `connect/3` for a single instance (standalone
mode). Both return an ered handle that can be used for sending commands.

`command/2,3,4` and `command_async/3,4` can be used for cluster and
standalone handles.

`close/1` is used for closing the connection(s).

Closes #27.

---------

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>