Move cluster API to ered_cluster module (#133)
The previous attempt #129 isn't good in practice. A client handle is
accessed by multiple processes, so it's very useful if it's a pid. Then
it can be registered with a name and clients from different processes
can use it.

In this attempt, the client handle and the cluster handle are just pids,
but the cluster API is moved to `ered_cluster` while the single node API
stays in `ered`.

---------

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>