Convergence check loop minimal fix
This patch fixes the convergence check busy loop problem by letting the
check time out rather than aborting it and restarting it immediately.

Also, server error replies like LOADING are skipped so we don't crash
on these in the convergence check.

This is a minimal rewrite of the convergence check fix:
https://github.com/Ericsson/ered/pull/146

Signed-off-by: Viktor Söderqvist <viktor.soderqvist@est.tech>